+++ malachite-test/tests/integer/arithmetic/shl_i32.rs
@@ -1618,7 +1618,7 @@ fn shl_round_i32_properties() {
         log_pairs(exhaustive_integers(), exhaustive_i::<i32>()),
         exhaustive_rounding_modes(),
     ).filter(|&((ref n, i), rm)| {
-        i >= 0 || rm != RoundingMode::Exact || n.divisible_by_power_of_2(i.wrapping_abs() as u64)
+        i >= 0 || rm != RoundingMode::Exact || n.divisible_by_power_of_2(i.wrapping_abs() as u32)
     })
         .take(LARGE_LIMIT)
     {
@@ -1631,7 +1631,7 @@ fn shl_round_i32_properties() {
         &(|seed| i32s_geometric(seed, 32)),
         &(|seed| random_rounding_modes(seed)),
     ).filter(|&(ref n, i, rm)| {
-        i >= 0 || rm != RoundingMode::Exact || n.divisible_by_power_of_2(i.wrapping_abs() as u64)
+        i >= 0 || rm != RoundingMode::Exact || n.divisible_by_power_of_2(i.wrapping_abs() as u32)
     })
         .take(LARGE_LIMIT)
     {
