+++ malachite-test/tests/natural/arithmetic/shl_i32.rs
@@ -998,7 +998,7 @@ fn shl_round_u32_properties() {
         log_pairs(exhaustive_naturals(), exhaustive_i::<i32>()),
         exhaustive_rounding_modes(),
     ).filter(|&((ref n, i), rm)| {
-        i >= 0 || rm != RoundingMode::Exact || n.divisible_by_power_of_2(-i as u64)
+        i >= 0 || rm != RoundingMode::Exact || n.divisible_by_power_of_2(-i as u32)
     })
         .take(LARGE_LIMIT)
     {
@@ -1011,7 +1011,7 @@ fn shl_round_u32_properties() {
         &(|seed| i32s_geometric(seed, 32)),
         &(|seed| random_rounding_modes(seed)),
     ).filter(|&(ref n, i, rm)| {
-        i >= 0 || rm != RoundingMode::Exact || n.divisible_by_power_of_2(-i as u64)
+        i >= 0 || rm != RoundingMode::Exact || n.divisible_by_power_of_2(-i as u32)
     })
         .take(LARGE_LIMIT)
     {
